backend/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
markdown==3.4.4
django-filter==23.2
django-cors-headers==4.2.0
pillow==10.0.0
python-dateutil==2.8.2
django-dbbackup==4.2.1
django-ninja==1.4.3
python-decouple==3.8
django-q2==1.6.2
django-jazzmin==3.0.0
django-import-export==4.0.3
arrow==1.3.0
pytz==2024.1
django-redis==6.0.0
pytest==9.0.2
pytest-django==4.11.1
django-model-utils==5.0.0
pytest-mock==3.15.1